+++ src/FromTask.ts
@@ -11,7 +11,7 @@ import { Task } from './Task'
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask<M> extends FromIO<M> {
+export interface FromTask<M> extends FromIO<M> {
   readonly fromTask: <A>(fa: Task<A>) => HKT<M, A>
 }
 
@@ -19,7 +19,7 @@ export interface MonadTask<M> extends FromIO<M> {
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask1<M extends URIS> extends FromIO1<M> {
+export interface FromTask1<M extends URIS> extends FromIO1<M> {
   readonly fromTask: <A>(fa: Task<A>) => Kind<M, A>
 }
 
@@ -27,7 +27,7 @@ export interface MonadTask1<M extends URIS> extends FromIO1<M> {
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask2<M extends URIS2> extends FromIO2<M> {
+export interface FromTask2<M extends URIS2> extends FromIO2<M> {
   readonly fromTask: <E, A>(fa: Task<A>) => Kind2<M, E, A>
 }
 
@@ -35,7 +35,7 @@ export interface MonadTask2<M extends URIS2> extends FromIO2<M> {
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask2C<M extends URIS2, E> extends FromIO2C<M, E> {
+export interface FromTask2C<M extends URIS2, E> extends FromIO2C<M, E> {
   readonly fromTask: <A>(fa: Task<A>) => Kind2<M, E, A>
 }
 
@@ -43,7 +43,7 @@ export interface MonadTask2C<M extends URIS2, E> extends FromIO2C<M, E> {
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask3<M extends URIS3> extends FromIO3<M> {
+export interface FromTask3<M extends URIS3> extends FromIO3<M> {
   readonly fromTask: <R, E, A>(fa: Task<A>) => Kind3<M, R, E, A>
 }
 
@@ -51,7 +51,7 @@ export interface MonadTask3<M extends URIS3> extends FromIO3<M> {
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask3C<M extends URIS3, E> extends FromIO3C<M, E> {
+export interface FromTask3C<M extends URIS3, E> extends FromIO3C<M, E> {
   readonly fromTask: <R, A>(fa: Task<A>) => Kind3<M, R, E, A>
 }
 
@@ -59,6 +59,6 @@ export interface MonadTask3C<M extends URIS3, E> extends FromIO3C<M, E> {
  * @category type classes
  * @since 3.0.0
  */
-export interface MonadTask4<M extends URIS4> extends FromIO4<M> {
+export interface FromTask4<M extends URIS4> extends FromIO4<M> {
   readonly fromTask: <S, R, E, A>(fa: Task<A>) => Kind4<M, S, R, E, A>
 }