+++ src/TaskEither.ts
@@ -38,7 +38,7 @@ import { IO } from './IO'
 import { IOEither } from './IOEither'
 import { bind_, chainFirst_, Monad2 } from './Monad'
 import { FromIO2 } from './FromIO'
-import { MonadTask2 } from './MonadTask'
+import { FromTask2 } from './FromTask'
 import { Monoid } from './Monoid'
 import { getLeft, getRight, Option } from './Option'
 import { Pointed2 } from './Pointed'
@@ -465,7 +465,7 @@ export const fromIO: FromIO2<URI>['fromIO'] = rightIO
  * @category MonadTask
  * @since 3.0.0
  */
-export const fromTask: MonadTask2<URI>['fromTask'] = rightTask
+export const fromTask: FromTask2<URI>['fromTask'] = rightTask
 
 // -------------------------------------------------------------------------------------
 // instances
@@ -773,7 +773,7 @@ export const FromIO: FromIO2<URI> = {
  * @category instances
  * @since 3.0.0
  */
-export const MonadTask: MonadTask2<URI> = {
+export const FromTask: FromTask2<URI> = {
   URI,
   fromIO,
   fromTask