+++ src/TaskThese.ts
@@ -12,7 +12,7 @@ import { IO } from './IO'
 import { IOEither } from './IOEither'
 import { Monad2C } from './Monad'
 import { FromIO2 } from './FromIO'
-import { MonadTask2 } from './MonadTask'
+import { FromTask2 } from './FromTask'
 import { Option } from './Option'
 import { Pointed2 } from './Pointed'
 import { Semigroup } from './Semigroup'
@@ -185,7 +185,7 @@ export const fromIO: FromIO2<URI>['fromIO'] = rightIO
  * @category MonadIO
  * @since 3.0.0
  */
-export const fromTask: MonadTask2<URI>['fromTask'] = rightTask
+export const fromTask: FromTask2<URI>['fromTask'] = rightTask
 
 // -------------------------------------------------------------------------------------
 // instances
@@ -341,7 +341,7 @@ export const FromIO: FromIO2<URI> = {
  * @category instances
  * @since 3.0.0
  */
-export const MonadTask: MonadTask2<URI> = {
+export const FromTask: FromTask2<URI> = {
   URI,
   fromIO,
   fromTask